# api/openapi.yaml
openapi: 3.1.0
info:
  title: omni-5e
  description: REST API for the D&D 5e System Reference Document
  version: 0.1.0
  license:
    name: CC-BY-4.0
    url: https://creativecommons.org/licenses/by/4.0/legalcode
  contact:
    name: omni-5e
    url: https://github.com/omni-5e/omni-5e
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
paths:
...
  /monsters:
    get:
      summary: List monsters
      operationId: listMonsters
      tags: [monsters]
      parameters:
        - $ref: "#/components/parameters/srd_version"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: cr_min
          in: query
          schema: { type: number }
        - name: cr_max
          in: query
          schema: { type: number }
        - name: type
          in: query
          schema: { type: string }
        - name: size
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string, enum: [monster, animal] }
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated list of monsters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonsterList"
  /monsters/{slug}:
    get:
      summary: Get a monster by slug
      operationId: getMonster
      tags: [monsters]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Monster detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Monster"
        "404":
          $ref: "#/components/responses/NotFound"
  /classes:
    get:
      summary: List classes
      operationId: listClasses
      tags: [classes]
      parameters:
        - $ref: "#/components/parameters/srd_version"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated list of classes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassList"
  /classes/{slug}:
    get:
      summary: Get a class by slug
      operationId: getClass
      tags: [classes]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Class detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Class"
        "404":
          $ref: "#/components/responses/NotFound"
  /classes/{slug}/subclasses:
    get:
      summary: List subclasses for a class
      operationId: getSubclasses
      tags: [classes]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Subclasses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subclass"
  /classes/{slug}/levels/{level}:
    get:
      summary: Get class level table row
      operationId: getClassLevel
      tags: [classes]
      parameters:
        - $ref: "#/components/parameters/slug"
        - name: level
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 20 }
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Level table row
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ClassLevelTableRow"
...
components:
  parameters:
    srd_version:
      name: srd_version
      in: query
      schema: { type: string }
      description: SRD content version (e.g. 5.2.1). Also via X-SRD-Version header.
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
    offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
    slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
    ProblemDetail:
      type: object
      properties:
        type: { type: string, example: "about:blank" }
        title: { type: string, example: "Not Found" }
        status: { type: integer, example: 404 }
        detail: { type: string }
...
    PaginationMeta:
      type: object
      properties:
        srd_version: { type: string }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    PaginationLinks:
      type: object
      properties:
        self: { type: string }
        next: { type: string, nullable: true }
        prev: { type: string, nullable: true }
...
    Monster:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        size: { type: string }
        type: { type: string }
        alignment: { type: string }
        ac:
          type: object
          properties:
            value: { type: integer }
            source: { type: string }
        hp:
          type: object
          properties:
            average: { type: integer }
            formula: { type: string }
        speed: { type: object, additionalProperties: { type: integer } }
        ability_scores:
          type: object
          properties:
            str: { type: integer }
            dex: { type: integer }
            con: { type: integer }
            int: { type: integer }
            wis: { type: integer }
            cha: { type: integer }
        cr: { type: number }
        xp: { type: integer }
        category: { type: string, enum: [monster, animal] }
        traits: { type: array, items: { $ref: "#/components/schemas/NamedBlock" } }
        actions: { type: array, items: { $ref: "#/components/schemas/NamedBlock" } }
        bonus_actions: { type: array, items: { $ref: "#/components/schemas/NamedBlock" } }
        reactions: { type: array, items: { $ref: "#/components/schemas/NamedBlock" } }
        legendary_actions: { type: array, items: { $ref: "#/components/schemas/NamedBlock" } }
    MonsterList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Monster" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
    NamedBlock:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
    Class:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        hit_die: { type: integer }
        primary_ability: { type: string }
        saving_throw_proficiencies: { type: array, items: { type: string } }
        armor_proficiencies: { type: array, items: { type: string } }
        weapon_proficiencies: { type: array, items: { type: string } }
        description: { type: string }
        subclasses: { type: array, items: { $ref: "#/components/schemas/Subclass" } }
    ClassList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Class" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
    Subclass:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        class_slug: { type: string }
        description: { type: string }
    ClassLevelTableRow:
      type: object
      properties:
        class_slug: { type: string }
        level: { type: integer }
        proficiency_bonus: { type: integer }
        features_unlocked: { type: array, items: { type: string } }
        other_columns: { type: object, additionalProperties: { type: string } }
...